% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{create_location_efs}
\alias{create_location_efs}
\title{Creates an endpoint for an Amazon EFS file system}
\usage{
create_location_efs(
  Subdirectory = "string",
  EfsFilesystemArn = "string",
  Ec2Config = list(
    SubnetArn = "string",
    SecurityGroupArns = list(
      "string"
    )
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{Subdirectory}{[required] A subdirectory in the location's path. This subdirectory in the EFS file system is used to read data from the EFS source location or write data to the EFS destination. By default, AWS DataSync uses the root directory.}

\item{EfsFilesystemArn}{[required] The Amazon Resource Name (ARN) for the Amazon EFS file system.}

\item{Ec2Config}{[required] The subnet and security group that the Amazon EFS file system uses.}

\item{Tags}{The key-value pair that represents a tag that you want to add to the resource. The value can be an empty string. This value helps you manage, filter, and search for your resources. We recommend that you create a name tag for your location.}
}
\description{
Creates an endpoint for an Amazon EFS file system.
}
\examples{

}
