% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{delete_task}
\alias{delete_task}
\title{Deletes a task}
\usage{
delete_task(TaskArn)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the task to delete.}
}
\description{
Deletes a task.
}
\section{Accepted Parameters}{
\preformatted{delete_task(
  TaskArn = "string"
)
}
}

