% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{describe_agent}
\alias{describe_agent}
\title{Returns metadata such as the name, the network interfaces, and the status (that is, whether the agent is running or not) for an agent}
\usage{
describe_agent(
  AgentArn = "string"
)
}
\arguments{
\item{AgentArn}{[required] The Amazon Resource Name (ARN) of the agent to describe.}
}
\description{
Returns metadata such as the name, the network interfaces, and the status (that is, whether the agent is running or not) for an agent. To specify which agent to describe, use the Amazon Resource Name (ARN) of the agent in your request.
}
\examples{

}
