% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{describe_location_efs}
\alias{describe_location_efs}
\title{Returns metadata, such as the path information about an Amazon EFS location}
\usage{
describe_location_efs(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the EFS location to describe.}
}
\description{
Returns metadata, such as the path information about an Amazon EFS location.
}
\section{Accepted Parameters}{
\preformatted{describe_location_efs(
  LocationArn = "string"
)
}
}

