% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{describe_location_s3}
\alias{describe_location_s3}
\title{Returns metadata, such as bucket name, about an Amazon S3 bucket location}
\usage{
describe_location_s3(
  LocationArn = "string"
)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the Amazon S3 bucket location to describe.}
}
\description{
Returns metadata, such as bucket name, about an Amazon S3 bucket location.
}
\examples{

}
