% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{describe_task}
\alias{describe_task}
\title{Returns metadata about a task}
\usage{
describe_task(
  TaskArn = "string"
)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the task to describe.}
}
\description{
Returns metadata about a task.
}
\examples{

}
