% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{describe_task_execution}
\alias{describe_task_execution}
\title{Returns detailed metadata about a task that is being executed}
\usage{
describe_task_execution(TaskExecutionArn)
}
\arguments{
\item{TaskExecutionArn}{[required] The Amazon Resource Name (ARN) of the task that is being executed.}
}
\description{
Returns detailed metadata about a task that is being executed.
}
\section{Accepted Parameters}{
\preformatted{describe_task_execution(
  TaskExecutionArn = "string"
)
}
}

