% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{list_agents}
\alias{list_agents}
\title{Returns a list of agents owned by an AWS account in the AWS Region specified in the request}
\usage{
list_agents(
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{MaxResults}{The maximum number of agents to list.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next list of agents.}
}
\description{
Returns a list of agents owned by an AWS account in the AWS Region specified in the request. The returned list is ordered by agent Amazon Resource Name (ARN).
}
\details{
By default, this operation returns a maximum of 100 agents. This operation supports pagination that enables you to optionally reduce the number of agents returned in a response.

If you have more agents than are returned in a response (that is, the response returns only a truncated list of your agents), the response contains a marker that you can specify in your next request to fetch the next page of agents.
}
\examples{

}
