% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{list_locations}
\alias{list_locations}
\title{Returns a lists of source and destination locations}
\usage{
list_locations(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of locations to return.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next list of locations.}
}
\description{
Returns a lists of source and destination locations.
}
\details{
If you have more locations than are returned in a response (that is, the response returns only a truncated list of your agents), the response contains a token that you can specify in your next request to fetch the next page of locations.
}
\section{Accepted Parameters}{
\preformatted{list_locations(
  MaxResults = 123,
  NextToken = "string"
)
}
}

