% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Returns all the tags associated with a specified resources}
\usage{
list_tags_for_resource(
  ResourceArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource whose tags to list.}

\item{MaxResults}{The maximum number of locations to return.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next list of locations.}
}
\description{
Returns all the tags associated with a specified resources.
}
\examples{

}
