% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_package.R
\docType{package}
\name{paws.datasync-package}
\alias{paws.datasync}
\alias{paws.datasync-package}
\title{paws.datasync: AWS DataSync}
\description{
AWS DataSync is a managed data transfer service that makes it
simpler for you to automate moving data between on-premises storage
and Amazon Simple Storage Service (Amazon S3) or Amazon Elastic File
System (Amazon EFS).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
