% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Applies a key-value pair to an AWS resource}
\usage{
tag_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to apply the tag to.}

\item{Tags}{[required] The tags to apply.}
}
\description{
Applies a key-value pair to an AWS resource.
}
\examples{

}
