% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes a tag from an AWS resource}
\usage{
untag_resource(ResourceArn, Keys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to remove the tag from.}

\item{Keys}{[required] The keys in the key-value pair in the tag to remove.}
}
\description{
Removes a tag from an AWS resource.
}
\section{Accepted Parameters}{
\preformatted{untag_resource(
  ResourceArn = "string",
  Keys = list(
    "string"
  )
)
}
}

