% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{update_agent}
\alias{update_agent}
\title{Updates the name of an agent}
\usage{
update_agent(AgentArn, Name = NULL)
}
\arguments{
\item{AgentArn}{[required] The Amazon Resource Name (ARN) of the agent to update.}

\item{Name}{The name that you want to use to configure the agent.}
}
\description{
Updates the name of an agent.
}
\section{Accepted Parameters}{
\preformatted{update_agent(
  AgentArn = "string",
  Name = "string"
)
}
}

