% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{update_task}
\alias{update_task}
\title{Updates the metadata associated with a task}
\usage{
update_task(TaskArn, Options = NULL, Name = NULL)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the resource name of the task to update.}

\item{Options}{}

\item{Name}{The name of the task to update.}
}
\description{
Updates the metadata associated with a task.
}
\section{Accepted Parameters}{
\preformatted{update_task(
  TaskArn = "string",
  Options = list(
    VerifyMode = "POINT_IN_TIME_CONSISTENT"|"NONE",
    Atime = "NONE"|"BEST_EFFORT",
    Mtime = "NONE"|"PRESERVE",
    Uid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    Gid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    PreserveDeletedFiles = "PRESERVE"|"REMOVE",
    PreserveDevices = "NONE"|"PRESERVE",
    PosixPermissions = "NONE"|"BEST_EFFORT"|"PRESERVE",
    BytesPerSecond = 123
  ),
  Name = "string"
)
}
}

