% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{delete_location}
\alias{delete_location}
\title{Deletes the configuration of a location used by AWS DataSync}
\usage{
delete_location(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the location to delete.}
}
\description{
Deletes the configuration of a location used by AWS DataSync.
}
\section{Accepted Parameters}{
\preformatted{delete_location(
  LocationArn = "string"
)
}
}

