% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{describe_location_nfs}
\alias{describe_location_nfs}
\title{Returns metadata, such as the path information, about a NFS location}
\usage{
describe_location_nfs(
  LocationArn = "string"
)
}
\arguments{
\item{LocationArn}{[required] The Amazon resource Name (ARN) of the NFS location to describe.}
}
\description{
Returns metadata, such as the path information, about a NFS location.
}
\examples{

}
