% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.datasync_operations.R
\name{list_tasks}
\alias{list_tasks}
\title{Returns a list of all the tasks}
\usage{
list_tasks(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of tasks to return.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next list of tasks.}
}
\description{
Returns a list of all the tasks.
}
\section{Accepted Parameters}{
\preformatted{list_tasks(
  MaxResults = 123,
  NextToken = "string"
)
}
}

